+++ apps/app-frontend/src/pages/instance/admonitions/shared-instance-unavailable.vue
