+++ apps/app-frontend/src/pages/instance/admonitions/shared-instance-update-available.vue
