+++ apps/app-frontend/src/pages/instance/page-header/instance-page-header-server-metadata.vue
